+++ packages/nestjs-org/src/interfaces/org-options-extras.interface.ts
@@ -1,9 +1,9 @@
 import { DynamicModule } from '@nestjs/common';
-import { OrgCrudBuilder } from '../org.crud-builder';
 import { OrgEntitiesOptionsInterface } from './org-entities-options.interface';
+import { OrgCrudBuilder } from '../utils/org.crud-builder';
 
 export interface OrgOptionsExtrasInterface
   extends Pick<DynamicModule, 'global' | 'controllers'>,
     Partial<OrgEntitiesOptionsInterface> {
-  crudBuilder?: OrgCrudBuilder;
+  orgCrudBuilder?: OrgCrudBuilder;
 }
